requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.13.0
azure-servicebus==7.11.1
azure-storage-blob==12.16.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
fastapi==0.99.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
isodate==0.6.1
msal==1.28.1
msal-extensions==1.1.0
multidict==6.0.5
packaging==24.1
portalocker==2.8.2
pycparser==2.22
pydantic==1.10.16
PyJWT==2.7.0
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2023.3
pywin32==306
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.22.0
yarl==1.9.4